@@ -11,7 +11,7 @@ class AppleSignInButton extends StatefulWidget {
   /// The callback that is called when the button is tapped or otherwise activated.
   ///
   /// If this is set to null, the button will be disabled.
-  final VoidCallback onPressed;
+  final VoidCallback? onPressed;
 
   /// A type for the authorization button.
   final ButtonType type;
@@ -51,7 +51,7 @@ class _AppleSignInButtonState extends State<AppleSignInButton> {
       onTapDown: (_) => setState(() => _isTapDown = true),
       onTapUp: (_) {
         setState(() => _isTapDown = false);
-        widget?.onPressed();
+        widget?.onPressed!();
       },
       onTapCancel: () => setState(() => _isTapDown = false),
       child: AnimatedContainer(
@@ -107,7 +107,7 @@ class _AppleSignInButtonState extends State<AppleSignInButton> {
 class _AppleLogoPainter extends CustomPainter {
   final Color color;
 
-  _AppleLogoPainter({@required this.color});
+  _AppleLogoPainter({required this.color});
 
   @override
   void paint(Canvas canvas, Size size) {
